content/base/test/test_bug600468.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=600468
-->
<head>
  <title>Test for Bug 600468</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=600468">Mozilla Bug 600468</a>
<p id="display"></p>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 600468 **/

var attr = document.createAttribute("foo");
attr.value = "bar";
ok(attr.firstChild, "attr should have a first child");

attr.value = "";
ok(!attr.firstChild, "attr should not have a first child");

</script>
</pre>
</body>
</html>
